fix: extension make header and panel transition together
